Changed BitsRemaining to uint8_t to match _bitsRemaining

+++ Core/NES/NesTypes.h
@@ -341,7 +341,7 @@ struct ApuDmcState
 	uint16_t Period;
 	uint16_t Timer;
 	uint16_t BytesRemaining;
-	uint16_t BitsRemaining;
+	uint8_t BitsRemaining;
 
 	uint8_t OutputVolume;
 };
