+++ UI/Interop/ConsoleState/NesState.cs
@@ -318,7 +318,7 @@ public struct NesApuDmcState
 	public UInt16 Period;
 	public UInt16 Timer;
 	public UInt16 BytesRemaining;
-	public UInt16 BitsRemaining;
+	public byte BitsRemaining;
 
 	public byte OutputVolume;
 }
